requirements.txt
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648010766492/work
certifi==2021.5.30
charset-normalizer==2.0.12
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
fonttools==4.32.0
idna==3.3
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
kiwisolver==1.4.2
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1647441664166/work
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
munkres==1.1.4
numexpr @ file:///home/builder/ci_310/numexpr_1640791965184/work
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2020.1
requests==2.27.1
scikit-learn @ file:///home/builder/ci_310/scikit-learn_1641557398975/work
scipy @ file:///home/builder/ci_310/scipy_1641389939032/work
seaborn @ file:///tmp/build/80754af9/seaborn_1600553570093/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tornado @ file:///home/builder/ci_310/tornado_1640795392144/work
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
